geometry namespace: simplify
