+++ two_d_game_ai/geometry/__init__.py
@@ -1,1 +1,14 @@
 """Package containing geometric classes and functions."""
+
+from two_d_game_ai import Vector2
+
+CIRCLE_DEGREES = 360
+
+
+def point_in_or_on_circle(
+    point: Vector2,
+    circle_centre: Vector2,
+    circle_radius: float,
+) -> bool:
+    """Return True if the point is (inside or on) the circle, else False."""
+    return (point - circle_centre).length() <= circle_radius
