--- two_d_game_ai/geometry/utils.py
@@ -1,16 +1,0 @@
-"""Module containing basic Cartesian navigation functions wrapped around
-`Pygame.Vector2` library.
-"""
-
-from two_d_game_ai import Vector2
-
-CIRCLE_DEGREES = 360
-
-
-def point_in_or_on_circle(
-    point: Vector2,
-    circle_centre: Vector2,
-    circle_radius: float,
-) -> bool:
-    """Return True if the point is (inside or on) the circle, else False."""
-    return (point - circle_centre).length() <= circle_radius
