+++ two_d_game_ai/render/__init__.py
@@ -4,6 +4,8 @@ from __future__ import annotations
 
 from math import radians
 
+from two_d_game_ai.geometry import CIRCLE_DEGREES
+
 
 def _to_display_radians(bearing_deg: float) -> float:
     """Convert bearing (degrees) to Pygame-compatible angle (radians).
@@ -21,4 +23,4 @@ def _to_display_radians(bearing_deg: float) -> float:
         Pygame-compatible angle in radians CW from East
 
     """
-    return radians(-bearing_deg + 90)
+    return radians(-bearing_deg + CIRCLE_DEGREES / 4)
