+++ two_d_game_ai/render/bot_renderer.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 
 from two_d_game_ai import Vector2
 from two_d_game_ai.entities import Bot
-from two_d_game_ai.geometry.utils import point_in_or_on_circle
+from two_d_game_ai.geometry import point_in_or_on_circle
 from two_d_game_ai.render import colors
 from two_d_game_ai.render.generic_entity_renderer import _GenericEntityRenderer
 from two_d_game_ai.render.primitives import _circle, _scaled_circular_arc, _scaled_line
