+++ two_d_game_ai/render/movement_block_renderer.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from two_d_game_ai.entities import MovementBlock
-from two_d_game_ai.geometry.utils import point_in_or_on_circle
+from two_d_game_ai.geometry import point_in_or_on_circle
 from two_d_game_ai.render import colors
 from two_d_game_ai.render.generic_entity_renderer import _GenericEntityRenderer
 from two_d_game_ai.render.primitives import _scaled_circle
